+++ argnorm/normalizers.py
@@ -20,8 +20,9 @@ class BaseNormalizer:
     Inherit this class and customize subclass methods to implement the normalization of new databases/formats.
     """
 
-    def __init__(self, database=None) -> None:
+    def __init__(self, database=None, tool_version=None) -> None:
         self.database = database
+        self.tool_version = tool_version
 
     def run(self, input_file : str):
         """
@@ -101,13 +102,20 @@ class ResFinderNormalizer(BaseNormalizer):
 
 
 class AMRFinderPlusNormalizer(BaseNormalizer):
-    def __init__(self, database=None) -> None:
+    def __init__(self, database=None, tool_version=None) -> None:
         database = 'ncbi'
-        super().__init__(database)
+        super().__init__(database, tool_version)
 
     def get_input_ids(self, itable):
-        gene_identifier = 'Sequence name'
-        accession = 'Accession of closest sequence'
+        if '3.10.30' in self.tool_version:
+            accession = 'Accession of closest sequence'
+            gene_identifier = 'Sequence name'
+        elif '4.0.' in self.tool_version or 'v4' in self.tool_version:
+            accession = 'Closest reference accession'
+            gene_identifier = 'Element name'
+        else:
+            raise Exception('Unsupported AMRFinderPlus version. Please use 3.10.30 or 4.0.19.')
+
         return pd.Series(itable[accession] + '|' + itable[gene_identifier].str.replace(' ', '_'))
 
     def preprocess_ref_genes(self, ref_genes):